+++ antarest/launcher/service.py
@@ -60,7 +60,7 @@ from antarest.launcher.model import (
 from antarest.launcher.repository import JobResultRepository, SolverPresetsRepository
 from antarest.login.service import LoginService
 from antarest.login.utils import current_user_context, get_current_user, require_current_user
-from antarest.output.output_service import OutputService
+from antarest.output.service import OutputService
 from antarest.study.repository import AccessPermissions, StudyFilter
 from antarest.study.service import StudyService
 from antarest.study.storage.utils import assert_permission, extract_output_name, find_single_output_path
