+++ antarest/output/storage/file/storage.py
@@ -43,7 +43,7 @@ from antarest.launcher.adapters.abstractlauncher import SimulationLogs
 from antarest.launcher.model import LogType
 from antarest.output.aggregator_management import AggregatorManager
 from antarest.output.filestudy.file_output_utils import extract_variables_list
-from antarest.output.output_model import OutputVariablesList
+from antarest.output.model import OutputVariablesList
 from antarest.output.storage.file.repository import FileOutputRepository
 from antarest.output.storage.output_storage import (
     IOutputStorage,
