+++ antarest/output/variable_view/model.py
@@ -14,7 +14,7 @@ from typing import Annotated, Literal, Tuple, TypeAlias
 from pydantic import BaseModel, Field
 
 from antarest.core.exceptions import OutputVariablesViewError
-from antarest.output.output_model import OutputVariablesList
+from antarest.output.model import OutputVariablesList
 from antarest.output.utils import MCIndAreasQueryFile, MCIndLinksQueryFile, QueryFileType
