+++ antarest/study/adapters.py
@@ -13,7 +13,7 @@ from pathlib import Path
 
 from typing_extensions import override
 
-from antarest.output.output_service import OutputService
+from antarest.output.service import OutputService
 from antarest.output.storage.output_storage import OutputDetails, OutputMetadata
 from antarest.study.service import IOutputsAccess
