+++ tests/launcher/test_service.py
@@ -65,7 +65,7 @@ from antarest.launcher.service import (
 )
 from antarest.login.model import Identity
 from antarest.login.utils import current_user_context, get_current_user
-from antarest.output.output_service import OutputService
+from antarest.output.service import OutputService
 from antarest.study.model import STUDY_VERSION_8_8, STUDY_VERSION_9_2, OwnerInfo, Study, StudyMetadataDTO
 from antarest.study.repository import StudyMetadataRepository
 from antarest.study.service import StudyService
