+++ tests/output/storage/v2/test_v2_output_storage.py
@@ -25,7 +25,7 @@ from antarest.launcher.model import LogType
 from antarest.lfs.dir_lfs import DirLargeFileStorage
 from antarest.lfs.lfs import ILargeFileStorage
 from antarest.output.storage.v2.repository import OutputV2Repository
-from antarest.output.storage.v2.v2_output_storage import V2OutputStorage
+from antarest.output.storage.v2.storage import V2OutputStorage
 from antarest.study.model import MatrixFrequency, MatrixIndex, Study
 from antarest.study.repository import StudyMetadataRepository
