+++ tests/output/variable_view/test_model.py
@@ -12,7 +12,7 @@
 import pytest
 
 from antarest.core.exceptions import OutputVariablesViewError
-from antarest.output.output_model import OutputVariablesList
+from antarest.output.model import OutputVariablesList
 from antarest.output.variable_view.model import (
     AreaOutputId,
     LinkOutputId,
