+++ tests/storage/web/test_studies_bp.py
@@ -34,8 +34,8 @@ from antarest.core.model import PublicMode
 from antarest.core.roles import RoleType
 from antarest.core.utils.archives import ArchiveFormat
 from antarest.matrixstore.service import MatrixService
-from antarest.output.output_blueprint import create_output_routes
-from antarest.output.output_service import OutputService
+from antarest.output.routes import create_output_routes
+from antarest.output.service import OutputService
 from antarest.study.main import build_study_service
 from antarest.study.model import (
     DEFAULT_WORKSPACE_NAME,
